+++ src/SortableList/demos/provider.tsx
@@ -1,8 +1,13 @@
+/**
+ * title: Provider
+ * description: 为了方便用户在更高的上下文中通过 `useSortableList()` hook 获得组件实例，我们提供了 `SortableListProvider` 由用户控制 Provider 的作用范围。
+ * compact: true
+ */
 import { SortableList, SortableListProvider, useSortableList } from '@ant-design/pro-editor';
 import { Button, message } from 'antd';
+import { useTheme } from 'antd-style';
 import { useState } from 'react';
 import { Flexbox } from 'react-layout-kit';
-
 import ItemRender from './_ItemRender';
 import { INIT_VALUES, SchemaItem } from './data';
 
@@ -10,32 +15,42 @@ const App = () => {
   const [listData, setListData] = useState<SchemaItem[]>(INIT_VALUES);
 
   return (
-    <Flexbox padding={24}>
-      <SortableList<SchemaItem>
-        value={listData}
-        onChange={(data) => {
-          console.log('data', data);
-          setListData(data);
-        }}
-        renderContent={(item, index) => <ItemRender item={item} index={index} />}
-        SHOW_STORE_IN_DEVTOOLS // 用于显示 Redux Devtools
-      />
-    </Flexbox>
+    <SortableList<SchemaItem>
+      value={listData}
+      onChange={(data) => {
+        console.log('data', data);
+        setListData(data);
+      }}
+      renderContent={(item, index) => <ItemRender item={item} index={index} />}
+      SHOW_STORE_IN_DEVTOOLS // 用于显示 Redux Devtools
+    />
   );
 };
 
 const Extra = () => {
   const instance = useSortableList();
   return (
-    <div>
-      <Button onClick={() => message.info(JSON.stringify(instance.getValue()))}>获取值</Button>
-    </div>
+    <Button
+      block
+      size="small"
+      type="primary"
+      style={{ marginBottom: 8 }}
+      onClick={() => message.info(JSON.stringify(instance.getValue()))}
+    >
+      获取值
+    </Button>
   );
 };
 
-export default () => (
-  <SortableListProvider>
-    <Extra />
-    <App />
-  </SortableListProvider>
-);
+export default () => {
+  const token = useTheme();
+
+  return (
+    <SortableListProvider>
+      <Flexbox padding={24} style={{ background: token.colorBgLayout }}>
+        <Extra />
+        <App />
+      </Flexbox>
+    </SortableListProvider>
+  );
+};