+++ src/SortableList/demos/renderContent.tsx
@@ -1,14 +1,18 @@
 /**
  * title: 自定义列表项内容
  * description: 提供 `renderContent` 由用户自定义除拖拽等操作外的列表项内容。
+ * compact: true
  */
 import { SortableList } from '@ant-design/pro-editor';
+import { useTheme } from 'antd-style';
+import { Flexbox } from 'react-layout-kit';
 import ItemRender from './_ItemRender';
 import { INIT_VALUES, SchemaItem } from './data';
 
 export default () => {
+  const token = useTheme();
   return (
-    <div style={{ width: 340, padding: '0 12px' }}>
+    <Flexbox padding={24} style={{ background: token.colorBgLayout }}>
       <SortableList<SchemaItem>
         initialValues={INIT_VALUES}
         onChange={(data, event) => {
@@ -16,6 +20,6 @@ export default () => {
         }}
         renderContent={(item, index) => <ItemRender item={item} index={index} />}
       />
-    </div>
+    </Flexbox>
   );
 };