+++ src/SortableList/type/store.ts
@@ -38,6 +38,10 @@ export interface SortableListState {
    */
   actions?: RenderActionProps;
   /**
+   * 获取唯一标识
+   */
+  getId?: (item: any, index: number) => UniqueIdentifier;
+  /**
    * 渲染可排序项
    */
   renderItem?: RenderItem;
@@ -85,6 +89,10 @@ export interface StoreUpdaterProps<T = SortableItem> {
    * 渲染可排序项
    */
   renderItem?: RenderItem<T>;
+  /**
+   * 获取唯一标识
+   */
+  getId?: (item: T, index: number) => UniqueIdentifier;
   /**
    * 渲染内容区域
    */